Extract animation logic from main.js (Issue #18)
